Update README.md
The target is given to the player and player starts with score of 0. 
Player chooses a number from 1 to 6 to reach towards the target and the code also generates a number randomly if the generated number is equal to the number entered by the user the game ends else if the numbers are not equal then the score of player increases by the the number entered. Once player's score equals or becomes greater than the target the user wins.